+++ ChangeLog.md
@@ -7,4 +7,5 @@ phpcpd 2.0.0
 ------------
 
 * Added the `--git-repository` switch to calculate the software metrics for each revision of a Git repository.
+* Removed the ability to count the Executable Lines of Code (ELOC) and the optional dependency on Bytekit.
 * The [Version](http://github.com/sebastianbergmann/version) component is now used to manage the version number.